services/MediaService/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.10.5
cffi==2.0.0
grpcio==1.76.0
grpcio-tools==1.76.0
grpcio-reflection==1.76.0
minio==7.2.18
protobuf==6.33.0
pycparser==2.23
pycryptodome==3.23.0
setuptools==80.9.0
typing_extensions==4.15.0
urllib3==2.5.0